+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/display/MiniAppDisplayViewModel.kt
@@ -430,7 +430,8 @@ class MiniAppDisplayViewModel constructor(
             ),
             miniAppId = appId,
             miniAppVersion = versionId,
-            fromCache = false
+            fromCache = false,
+            fromBundle = true
         )
     }