+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## CHANGELOG
 
+### 5.5.0 (2023-10-05)
+**SDK**
+- **Fix:** Improve loading MiniApp from bundle loading time by removing few checks
+
+### 5.4.2 (2023-08-29)
+**SDK**
+- **Fix:** Store hash on successfull download of Miniapp in background.
+
 ### 5.4.1 (2023-08-29)
 **SDK**
 - **Fix:** Prevent manifest file being deleted when `MiniAppView.loadFromBundle` being called.